Fix support for ovpn files
This fixes the support for ovpn files.  The prior method would never work, because echo *.{conf,ovpn) always results in two entries. *.conf or the resolved conf if one or more files exist, and *.ovpn or the resolve ovpn files if one or more exist.  This version resolves the issue by listing all files in the directory, and egrep filters the results to all *.conf or *.ovpn files.  If there is just one (total) then it is assigned.  The same was done for the crt/cert files.