Fix initialization/configuration failures when using oven file
because of the nature of the ls command with wildcards, configurations like mine that use the .ovpn file fail to work unless I also create the conf file.  This fix allows the ovpn file to be used without a complementing conf file.  

The way the wildcard parser works, the logic to override the conf file fails, because the ls command results in at least 1 file not found, and thus the && true conditional execution event cannot run.